dependsOn() update
spec/Spec.js depends on lib/a.js and /lib.bjs

if lib/a.js is modified spec/Spec.js was always succesfully returned.

But depends() would return an empty array if spec/Spec.js was modified, since nothing depends on it.

Now, if spec/Spec.js is modified and it's one of the files that we created dependency list of, dependsOn() will return spec/Spec.js itself too.